Refactor UI layouts and improve code structure
- Updated the dimensions and positions of widgets in order_window.ui and write_card_loading.ui for better alignment and aesthetics.
- Removed unnecessary QLabel widgets and replaced them with QHBoxLayout for improved layout management.
- Simplified the construction of WriteCard and ClearCard objects by using parameterized constructors.
- Cleaned up the MainWindow destructor and removed commented-out code for clarity.
- Enhanced the order information extraction logic in Order class to handle project names more robustly.
- Removed unused includes and improved variable types for better readability and performance.
- Updated loading window titles for better user feedback during operations.
- Adjusted font sizes in UI elements for consistency and improved readability.
- Cleaned up the upload_file and reset_card classes by removing commented-out code.

+++ src/order/person_data.cpp
@@ -1,7 +1,6 @@
 #include "person_data.h"
 
 #include <memory>
-#include <iostream>
 
 using namespace zel::json;
 using namespace zel::filesystem;
@@ -23,7 +22,7 @@ std::shared_ptr<PersonDataInfo> PersonData::personDataInfo() {
 
     auto person_data_file_path_pgp = person_data_files->at(0).path();
     if (person_data_file_path_pgp.find(".gpg") != std::string::npos || person_data_file_path_pgp.find(".pgp") != std::string::npos) {
-        int pos = person_data_file_path_pgp.find_last_of(".");
+        size_t pos = person_data_file_path_pgp.find_last_of(".");
         if (pos == std::string::npos) return nullptr;
         auto person_data_file_path = person_data_file_path_pgp.substr(0, pos);
 
@@ -64,7 +63,7 @@ zel::json::Json PersonData::jsonData() {
 
     auto headers = String::split(person_data_info_->header, "/");
     auto datas   = String::split(person_data_info_->data, " ");
-    for (int i = 0; i < headers.size(); i++) {
+    for (size_t i = 0; i < headers.size(); i++) {
         json[headers[i]] = datas[i];
     }
